Finish PortStreamWriter
