+++ src/main/scala/Port.scala
@@ -22,6 +22,7 @@ case class PortIOConfig(
 ) {
   def nAddrBits: Int = if (nElems > 1) log2Ceil(nElems) else 1
 }
+
 // PortIn is a antiDependency
 class PortInIO[T <: Data](c: PortIOConfig, gen: T) extends Bundle {
   val en = Input(Bool())
@@ -103,84 +104,4 @@ class RegPort[T <: Data](c: PortConfig, gen: T) extends Port[T](c, gen) {
   }
 }
 
-class PortStreamReaderIO[T <: Data](c: PortIOConfig, gen: T) extends Bundle {
-  val portRead = Flipped(new PortOutIO(c,gen))
-  val start = Flipped(Decoupled())
-  val done = Output(Bool())
 
-  val out = Decoupled(gen)
-
-  def tieOff(): Unit = {
-    portRead.addr := 0.U
-    portRead.en := false.B
-    start.ready := false.B
-    done := false.B
-    out.valid := false.B
-    out.bits := 0.U.asTypeOf(gen)
-  }
-}
-class PortStreamReader[T <: Data](c: PortIOConfig, gen: T) extends Module {
-  val io = IO(new PortStreamReaderIO(c,gen))
-  io.tieOff()
-
-  val regCnt = RegInit(0.U((c.nAddrBits+1).W))
-  val regRspValid = RegInit(false.B)
-  val wReadEn = WireInit(false.B)
-  wReadEn := false.B
-
-  regRspValid := wReadEn
-
-  val sIdle :: sRunning :: sDone :: Nil = Enum(3)
-  val queueSize = 4
-  val queue = Module(new Queue(gen, queueSize, useSyncReadMem = true))
-  queue.io.deq <> io.out
-  queue.io.enq.bits := io.portRead.data
-  queue.io.enq.valid := regRspValid
-  assert(!(regRspValid && !queue.io.enq.ready), "[Port.scala] PortStreamReader read data but queue was not ready")
-
-  val regState = RegInit(sIdle)
-
-  switch(regState) {
-    is (sIdle) {
-      io.start.ready := true.B
-      when (io.start.fire) {
-        regState := sRunning
-        regCnt := 0.U
-      }
-    }
-
-    is (sRunning) {
-      when(regCnt === (c.nElems).U) {
-        regState := sDone
-        assert(queue.io.enq.fire, "[Port.scala] PortStreamReader should have enqueued last value")
-      } otherwise {
-        // Guarantee that we will have room for the data when it arrives next cycle
-        when(queue.io.count < (queueSize - 2).U) {
-          io.portRead.en := true.B
-          wReadEn := true.B
-          io.portRead.addr := regCnt
-
-          regCnt := regCnt + 1.U
-        }
-      }
-    }
-
-    is (sDone) {
-     when(queue.io.count === 0.U) {
-       io.done := true.B
-       regState := sIdle
-       regCnt := 0.U
-     }
-    }
-  }
-}
-
-class PortStreamWriterIO[T <: Data](c: PortIOConfig, gen: T) extends Bundle {
-  val in = Flipped(Decoupled(gen))
-  val portWrite = Flipped(new PortInIO(c,gen))
-  val start = Flipped(Decoupled())
-  val done = Output(Bool())
-}
-class PortStreamWriter[T <: Data](c: PortIOConfig, gen: T) extends Module {
-  val io = IO(new PortStreamWriterIO(c,gen))
-}
